fix: replace useSearchParams with server searchParams + clean 404
