+++ app/_not-found/page.tsx
@@ -1,0 +1,8 @@
+export default function NotFound() {
+  return (
+    <div className="p-6">
+      <h1 className="text-xl font-semibold">Page not found</h1>
+      <p className="text-sm opacity-80">The page you’re looking for doesn’t exist.</p>
+    </div>
+  );
+}
